redux-vs09 string filter moved to functions module
